Refactor infrastructure services - Deleted unused DateTimeService class - Deleted unused EmailService class - Modified EmailService class in Infrastructure/Services directory
